wip: more macros
