+++ rust/db_proto/src/arrays.rs
@@ -236,138 +236,6 @@ impl<'a, T: FieldAccessArray> Iterator for ArrayIter<'a, T> {
     }
 }
 
-/// Define array accesses for inflated, strongly-typed arrays of both
-/// zero-terminated and length-delimited types.
-#[macro_export]
-#[doc(hidden)]
-macro_rules! array_access {
-    ($acc:ident :: FieldAccess, $ty:ty) => {
-        $crate::array_access!($acc :: FieldAccess, $ty | u8 i16 i32 u32);
-    };
-    ($acc:ident :: FieldAccess, $ty:ty | $($len:ty)*) => {
-        $(
-        #[allow(unused)]
-        impl $acc::FieldAccess<$crate::meta::Array<$len, $ty>> {
-            pub const fn meta() -> &'static dyn $crate::Meta {
-                &$crate::meta::Array::<$len, $ty> { _phantom: std::marker::PhantomData }
-            }
-            #[inline]
-            pub const fn size_of_field_at(mut buf: &[u8]) -> Result<usize, $crate::ParseError> {
-                let mut size = std::mem::size_of::<$len>();
-                let mut len = match $acc::FieldAccess::<$len>::extract(buf) {
-                    Ok(n) => n,
-                    Err(e) => return Err(e),
-                };
-                #[allow(unused_comparisons)]
-                if len < 0 {
-                    return Err($crate::ParseError::InvalidData);
-                }
-                buf = buf.split_at(size).1;
-                loop {
-                    if len <= 0 {
-                        break;
-                    }
-                    len -= 1;
-                    let elem_size = match $acc::FieldAccess::<$ty>::size_of_field_at(buf) {
-                        Ok(n) => n,
-                        Err(e) => return Err(e),
-                    };
-                    buf = buf.split_at(elem_size).1;
-                    size += elem_size;
-                }
-                Ok(size)
-            }
-            #[inline(always)]
-            pub const fn extract(buf: &[u8]) -> Result<$crate::Array<'_, $len, $ty>, $crate::ParseError> {
-                match $acc::FieldAccess::<$len>::extract(buf) {
-                    Ok(len) => Ok($crate::Array::new(buf.split_at(std::mem::size_of::<$len>()).1, len as u32)),
-                    Err(e) => Err(e)
-                }
-            }
-            #[inline]
-            pub const fn measure<'a>(buffer: &'a[<$ty as $crate::Enliven>::ForMeasure<'a>]) -> usize {
-                let mut size = std::mem::size_of::<$len>();
-                let mut index = 0;
-                loop {
-                    if index + 1 > buffer.len() {
-                        break;
-                    }
-                    let item = &buffer[index];
-                    size += $acc::FieldAccess::<$ty>::measure(item);
-                    index += 1;
-                }
-                size
-            }
-            #[inline(always)]
-            pub fn copy_to_buf<'a>(buf: &mut $crate::BufWriter, value: &'a[<$ty as $crate::Enliven>::ForBuilder<'a>]) {
-                buf.write(&<$len>::to_be_bytes(value.len() as _));
-                for elem in value {
-                    $acc::FieldAccess::<$ty>::copy_to_buf(buf, elem);
-                }
-            }
-            #[inline(always)]
-            pub const fn constant(value: usize) -> $crate::Array<'static, $len, $ty> {
-                panic!("Constants unsupported for this data type")
-            }
-        }
-        )*
-
-        #[allow(unused)]
-        impl $acc::FieldAccess<$crate::meta::ZTArray<$ty>> {
-            pub const fn meta() -> &'static dyn $crate::Meta {
-                &$crate::meta::ZTArray::<$ty> { _phantom: std::marker::PhantomData }
-            }
-            #[inline]
-            pub const fn size_of_field_at(mut buf: &[u8]) -> Result<usize, $crate::ParseError> {
-                let mut size = 1;
-                loop {
-                    if buf.is_empty() {
-                        return Err($crate::ParseError::TooShort);
-                    }
-                    if buf[0] == 0 {
-                        return Ok(size);
-                    }
-                    let elem_size = match $acc::FieldAccess::<$ty>::size_of_field_at(buf) {
-                        Ok(n) => n,
-                        Err(e) => return Err(e),
-                    };
-                    buf = buf.split_at(elem_size).1;
-                    size += elem_size;
-                }
-            }
-            #[inline(always)]
-            pub const fn extract(mut buf: &[u8]) -> Result<$crate::ZTArray<$ty>, $crate::ParseError> {
-                Ok($crate::ZTArray::new(buf))
-            }
-            #[inline]
-            pub const fn measure<'a>(buffer: &'a[<$ty as $crate::Enliven>::ForMeasure<'a>]) -> usize {
-                let mut size = 1;
-                let mut index = 0;
-                loop {
-                    if index + 1 > buffer.len() {
-                        break;
-                    }
-                    let item = &buffer[index];
-                    size += $acc::FieldAccess::<$ty>::measure(item);
-                    index += 1;
-                }
-                size
-            }
-            #[inline(always)]
-            pub fn copy_to_buf(buf: &mut $crate::BufWriter, value: &[<$ty as $crate::Enliven>::ForBuilder<'_>]) {
-                for elem in value {
-                    $acc::FieldAccess::<$ty>::copy_to_buf(buf, elem);
-                }
-                buf.write_u8(0);
-            }
-            #[inline(always)]
-            pub const fn constant(value: usize) -> $crate::ZTArray<'static, $ty> {
-                panic!("Constants unsupported for this data type")
-            }
-        }
-    };
-}
-
 // Arrays of type [`u8`] are special-cased to return a slice of bytes.
 impl<T> AsRef<[u8]> for Array<'_, T, u8> {
     fn as_ref(&self) -> &[u8] {
