+++ rust/db_proto/src/datatypes.rs
@@ -567,87 +567,6 @@ macro_rules! basic_types {
 
     (: array<$ty:ty> $($len:ty)*) => {
         $(
-            #[allow(unused)]
-            impl FieldAccess<ArrayMeta<$len, $ty>> {
-                pub const fn meta() -> &'static dyn Meta {
-                    &ArrayMeta::<$len, $ty> { _phantom: PhantomData }
-                }
-                #[inline(always)]
-                pub const fn size_of_field_at(buf: &[u8]) -> Result<usize, $crate::ParseError> {
-                    const N: usize = std::mem::size_of::<$ty>();
-                    const L: usize = std::mem::size_of::<$len>();
-                    if let Some(len) = buf.first_chunk::<L>() {
-                        let len_value = <$len>::from_be_bytes(*len);
-                        #[allow(unused_comparisons)]
-                        if len_value < 0 {
-                            return Err($crate::ParseError::InvalidData);
-                        }
-                        let mut byte_len = len_value as usize;
-                        byte_len = match byte_len.checked_mul(N) {
-                            Some(l) => l,
-                            None => return Err($crate::ParseError::TooShort),
-                        };
-                        byte_len = match byte_len.checked_add(L) {
-                            Some(l) => l,
-                            None => return Err($crate::ParseError::TooShort),
-                        };
-                        if buf.len() < byte_len {
-                            Err($crate::ParseError::TooShort)
-                        } else {
-                            Ok(byte_len)
-                        }
-                    } else {
-                        Err($crate::ParseError::TooShort)
-                    }
-                }
-                #[inline(always)]
-                pub const fn extract(mut buf: &[u8]) -> Result<Array<$len, $ty>, $crate::ParseError> {
-                    const N: usize = std::mem::size_of::<$ty>();
-                    const L: usize = std::mem::size_of::<$len>();
-                    if let Some((len, array)) = buf.split_first_chunk::<L>() {
-                        let len_value = <$len>::from_be_bytes(*len);
-                        #[allow(unused_comparisons)]
-                        if len_value < 0 {
-                            return Err($crate::ParseError::InvalidData);
-                        }
-                        let mut byte_len = len_value as usize;
-                        byte_len = match byte_len.checked_mul(N) {
-                            Some(l) => l,
-                            None => return Err($crate::ParseError::TooShort),
-                        };
-                        byte_len = match byte_len.checked_add(L) {
-                            Some(l) => l,
-                            None => return Err($crate::ParseError::TooShort),
-                        };
-                        if buf.len() < byte_len {
-                            Err($crate::ParseError::TooShort)
-                        } else {
-                            Ok(Array::new(array, <$len>::from_be_bytes(*len) as u32))
-                        }
-                    } else {
-                        Err($crate::ParseError::TooShort)
-                    }
-                }
-                #[inline(always)]
-                pub const fn measure(buffer: &[$ty]) -> usize {
-                    buffer.len() * std::mem::size_of::<$ty>() + std::mem::size_of::<$len>()
-                }
-                #[inline(always)]
-                pub fn copy_to_buf(mut buf: &mut BufWriter, value: &[$ty]) {
-                    let size: usize = std::mem::size_of::<$ty>() * value.len() + std::mem::size_of::<$len>();
-                    if !buf.test(size) {
-                        return;
-                    }
-                    buf.write(&<$len>::to_be_bytes(value.len() as _));
-                    for n in value {
-                        buf.write(&<$ty>::to_be_bytes(*n));
-                    }
-                }
-                #[inline(always)]
-                pub const fn constant(value: usize) -> Array<'static, $len, $ty> {
-                    panic!("Constants unsupported for this data type")
-                }
-            }
         )*
     }
 }
